+++ src/types/utils.d.ts
@@ -1,22 +1,40 @@
-export interface EncryptAsymmetric {
+export interface IEncryptAsymmetricInput {
     plaintext: string;
     publicKey: string;
     privateKey: string;
 }
 
-export interface DecryptAsymmetric {
+export interface IDecryptAsymmetricInput {
     ciphertext: string;
     nonce: string;
     publicKey: string;
     privateKey: string;
 }
 
-export interface EncryptSymmetric {
+export interface IEncryptSymmetricInput {
     plaintext: string;
     key: string;
 }
 
-export interface DecryptSymmetric {
+export interface IDecryptSymmetricInput {
+	ciphertext: string;
+	iv: string;
+	tag: string;
+	key: string;
+}
+
+export interface IEncryptSymmetricOutput {
+    ciphertext: string;
+    iv: string;
+    tag: string;
+}
+
+export interface IEncryptSymmetric128BitHexKeyUTF8Input {
+    plaintext: string;
+    key: string;
+}
+
+export interface IDecryptSymmetric128BitHexKeyUTF8Input {
     ciphertext: string;
     iv: string;
     tag: string;