+++ src/utils/crypto.ts
@@ -1,7 +1,23 @@
 import * as crypto from 'crypto';
 import * as nacl from 'tweetnacl';
 import * as util from 'tweetnacl-util';
-import { DecryptAsymmetric, DecryptSymmetric, EncryptAsymmetric, EncryptSymmetric } from '../types/utils';
+import { 
+    IEncryptAsymmetricInput, 
+    IEncryptSymmetric128BitHexKeyUTF8Input,
+    IEncryptSymmetricInput,
+    IEncryptSymmetricOutput,
+    IDecryptAsymmetricInput, 
+    IDecryptSymmetric128BitHexKeyUTF8Input,
+    IDecryptSymmetricInput
+} from '../types/utils';
+import {
+    IV_BYTES_SIZE,
+    SYMMETRIC_KEY_BYTES_SIZE,
+    ALGORITHM_AES_256_GCM,
+    ENCODING_SCHEME_BASE64,
+    ENCODING_SCHEME_HEX,
+    ENCODING_SCHEME_UTF8
+} from '../variables';
 
 /**
  * Return assymmetrically encrypted [plaintext] using [publicKey] where
@@ -14,7 +30,7 @@ import { DecryptAsymmetric, DecryptSymmetric, EncryptAsymmetric, EncryptSymmetri
  * @returns {String} ciphertext - base64-encoded ciphertext
  * @returns {String} nonce - base64-encoded nonce
  */
-const encryptAsymmetric = ({ plaintext, publicKey, privateKey }: EncryptAsymmetric) => {
+export const encryptAsymmetric = ({ plaintext, publicKey, privateKey }: IEncryptAsymmetricInput) => {
 	const nonce = nacl.randomBytes(24);
     const ciphertext = nacl.box(
         util.decodeUTF8(plaintext),
@@ -39,7 +55,7 @@ const encryptAsymmetric = ({ plaintext, publicKey, privateKey }: EncryptAsymmetr
  * @param {String} obj.privateKey - private key of the receiver (current user)
  * @param {String} plaintext - UTF8 plaintext
  */
-const decryptAsymmetric = ({ ciphertext, nonce, publicKey, privateKey }: DecryptAsymmetric) => nacl.box.open(
+export const decryptAsymmetric = ({ ciphertext, nonce, publicKey, privateKey }: IDecryptAsymmetricInput) => nacl.box.open(
         util.decodeBase64(ciphertext),
         util.decodeBase64(nonce),
         util.decodeBase64(publicKey),
@@ -47,12 +63,94 @@ const decryptAsymmetric = ({ ciphertext, nonce, publicKey, privateKey }: Decrypt
 );
 
 /**
+ * Return new symmetric key with 
+ * @param {Object} options
+ * @param {Number} size - size of key to create in bits (128, 256, 512)
+ * @param {String} encoding - encoding of key to create (utf8, hex, base64)
+ * @returns {String} key - new symmetric key
+ */
+export const createSymmetricKey = (): string => 
+    crypto.randomBytes(SYMMETRIC_KEY_BYTES_SIZE).toString('base64');
+
+/**
  * Return symmetrically encrypted [plaintext] using [key].
  * @param {Object} obj
- * @param {String} obj.plaintext - plaintext to encrypt
- * @param {String} obj.key - hex key
+ * @param {String} obj.plaintext - (utf8) plaintext to encrypt
+ * @param {String} obj.key - (base64) 256-bit key
+ * @returns {Object} obj
+ * @returns {String} obj.ciphertext (base64) ciphertext
+ * @returns {String} obj.iv (base64) iv
+ * @returns {String} obj.tag (base64) tag
  */
-const encryptSymmetric = ({ plaintext, key }: EncryptSymmetric) => {
+export const encryptSymmetric = ({
+	plaintext,
+	key
+}: IEncryptSymmetricInput): IEncryptSymmetricOutput => {
+    
+    const iv = crypto.randomBytes(IV_BYTES_SIZE);
+
+    const secretKey = crypto.createSecretKey(key, 'base64');
+    const cipher = crypto.createCipheriv(ALGORITHM_AES_256_GCM, secretKey, iv);
+
+    let ciphertext = cipher.update(plaintext, 'utf8', 'base64');
+    ciphertext += cipher.final('base64');
+
+	return {
+		ciphertext,
+		iv: iv.toString('base64'),
+		tag: cipher.getAuthTag().toString('base64')
+	};
+};
+
+/**
+ * Return symmetrically decrypted [ciphertext] using [iv], [tag],
+ * and [key].
+ * @param {Object} obj
+ * @param {String} obj.ciphertext - ciphertext to decrypt
+ * @param {String} obj.iv - (base64) 256-bit iv
+ * @param {String} obj.tag - (base64) tag
+ * @param {String} obj.key - (base64) 256-bit key
+ * @returns {String} cleartext - the deciphered ciphertext
+ */
+export const decryptSymmetric = ({
+	ciphertext,
+	iv,
+	tag,
+	key
+}: IDecryptSymmetricInput): string => {
+
+    const secretKey = crypto.createSecretKey(key, 'base64');
+
+    const decipher = crypto.createDecipheriv(
+        ALGORITHM_AES_256_GCM,
+        secretKey,
+        Buffer.from(iv, 'base64')
+    );
+
+    decipher.setAuthTag(Buffer.from(tag, 'base64'));
+
+    let cleartext = decipher.update(ciphertext, 'base64', 'utf8');
+    cleartext += decipher.final('utf8');
+
+    return cleartext;
+};
+
+/**
+ * Return symmetrically encrypted [plaintext] using [key].
+ * 
+ * NOTE: THIS FUNCTION SHOULD NOT BE USED FOR ALL FUTURE
+ * ENCRYPTION OPERATIONS UNLESS IT TOUCHES OLD FUNCTIONALITY
+ * THAT USES IT.
+ * 
+ * @param {Object} obj
+ * @param {String} obj.plaintext - (utf8) plaintext to encrypt
+ * @param {String} obj.key - (hex) 128-bit key
+ * @returns {Object} obj
+ * @returns {String} obj.ciphertext (base64) ciphertext
+ * @returns {String} obj.iv (base64) iv
+ * @returns {String} obj.tag (base64) tag
+ */
+export const encryptSymmetric128BitHexKeyUTF8 = ({ plaintext, key }: IEncryptSymmetric128BitHexKeyUTF8Input) => {
     const ALGORITHM = 'aes-256-gcm';
     const BLOCK_SIZE_BYTES = 16;
 
@@ -69,16 +167,14 @@ const encryptSymmetric = ({ plaintext, key }: EncryptSymmetric) => {
 } 
 
 /**
- * Return symmetrically decypted [ciphertext] using [iv], [tag],
+ * Return symmetrically decrypted [ciphertext] using [iv], [tag],
  * and [key].
- * @param {Object} obj
- * @param {String} obj.ciphertext - ciphertext to decrypt
- * @param {String} obj.iv - iv
- * @param {String} obj.tag - tag
- * @param {String} obj.key - hex key
- *
- */
-const decryptSymmetric = ({ ciphertext, iv, tag, key }: DecryptSymmetric) => {
+ * 
+ * NOTE: THIS FUNCTION SHOULD NOT BE USED FOR ALL FUTURE
+ * DECRYPTION OPERATIONS UNLESS IT TOUCHES OLD FUNCTIONALITY
+ * THAT USES IT. USE decryptSymmetric() instead
+*/
+export const decryptSymmetric128BitHexKeyUTF8 = ({ ciphertext, iv, tag, key }: IDecryptSymmetric128BitHexKeyUTF8Input) => {
     const ALGORITHM = 'aes-256-gcm';
     
     const decipher = crypto.createDecipheriv(
@@ -93,10 +189,3 @@ const decryptSymmetric = ({ ciphertext, iv, tag, key }: DecryptSymmetric) => {
 
     return cleartext;
 }
-
-export {
-	encryptAsymmetric,
-	decryptAsymmetric,
-	encryptSymmetric,
-	decryptSymmetric
-};