+++ src/variables/index.ts
@@ -1,7 +1,9 @@
-const INFISICAL_URL = 'https://app.infisical.com';
-const AUTH_MODE_SERVICE_TOKEN = 'serviceToken';
+export const INFISICAL_URL = 'https://app.infisical.com';
+export const AUTH_MODE_SERVICE_TOKEN = 'serviceToken';
+export const ALGORITHM_AES_256_GCM = 'aes-256-gcm';
+export const IV_BYTES_SIZE = 12;
+export const SYMMETRIC_KEY_BYTES_SIZE = 32;
 
-export {
-    INFISICAL_URL,
-    AUTH_MODE_SERVICE_TOKEN
-}
+export const ENCODING_SCHEME_UTF8 = 'utf8';
+export const ENCODING_SCHEME_HEX = 'hex';
+export const ENCODING_SCHEME_BASE64 = 'base64';